+++ lib/src/edit.dart
@@ -17,19 +17,6 @@ class Edit {
   ///Given a prompt and an instruction,
   /// the model will return an edited
   /// version of the prompt.[prompt]
-  /**
-   * ### Example
-   * ```dart
-   * void editPrompt() async {
-      final response = await openAI.editor.prompt(EditRequest(
-      model: EditModel.TextEditModel,
-      input: 'What day of the wek is it?',
-      instruction: 'Fix the spelling mistakes'));
-
-      print(response.choices.last.text);
-      }
-   * ```
-   */
   Future<EditResponse> prompt(EditRequest request) {
     return _client.post(kURL + kEditPrompt, _cancel, request.toJson(),
         onSuccess: (it) {
@@ -39,22 +26,6 @@ class Edit {
 
   ///Creates an edited or extended image
   /// given an original image and a prompt.[editImage]
-  /**
-   * ### Example
-   * ```dart
-   *   void editImage() async {
-      final image = File('image-path');
-
-      final response = await openAI.editor.editImage(EditImageRequest(
-      image: EditFile("${image?.path}", '${image?.name}'),
-      mask: EditFile('file path', 'file name'),
-      size: ImageSize.size1024,
-      prompt: 'King Snake'));
-
-      print(response.data?.last?.url);
-      }
-   * ```
-   */
   Future<GenImgResponse> editImage(EditImageRequest request) async {
     final mRequest = await request.convert();
     return _client.postFormData(kURL + kImageEdit, _cancel, mRequest,
@@ -64,20 +35,6 @@ class Edit {
   }
 
   ///Creates a variation of a given image.[variation]
-  /**
-   * ### example
-   * ```dart
-   *  void onVariation() async {
-     final image = file('image-path');
-
-      final request =
-      Variation(image: EditFile('${image?.path}', '${image?.name}'));
-      final response = await openAI.editor.variation(request);
-
-      print(response.data?.last?.url);
-      }
-   * ```
-   */
   Future<GenImgResponse> variation(Variation request) async {
     final mRequest = await request.convert();
     return _client.postFormData(kURL + kVariation, _cancel, mRequest,