+++ lib/src/model/fine_tune/response/fine_tune_delete.dart
@@ -1,0 +1,11 @@
+class FineTuneDelete {
+  final String id;
+  final String object;
+  final bool deleted;
+
+  FineTuneDelete(
+      {required this.id, required this.object, required this.deleted});
+
+  factory FineTuneDelete.fromJson(Map<String, dynamic> json) => FineTuneDelete(
+      id: json['id'], object: json['object'], deleted: json['deleted']);
+}